media menu fixed
