+++ styles/Modal.module.css
@@ -15,9 +15,10 @@
 }
 
 .userSettings{
-  width: 550px;
+  width: 35rem;
+  height: 22rem;
 }
 
 .mediaSelect{
-  width: 8rem;
+  width: 11rem;
 }
